flytt funksjonen tilBarnasSkalIkkeUtbetalesTidslinjer inn i UtbetalingTidslinjeService og flytt tester
